+++ lib/src/widgets/reactive_form_pop_scope.dart
@@ -14,24 +14,18 @@ class ReactiveFormPopScope<T> extends StatelessWidget {
   final ReactiveFormCanPopCallback? canPop;
 
   /// A callback invoked when a route is popped. See [PopScope] for more details.
-  final ReactiveFormPopInvokedCallback? onPopInvoked;
-
-  /// A callback invoked when a route is popped. See [PopScope] for more details.
   final ReactiveFormPopInvokedWithResultCallback<T>? onPopInvokedWithResult;
 
   const ReactiveFormPopScope({
     super.key,
     this.canPop,
-    @Deprecated('Use onPopInvokedWithResult instead.') this.onPopInvoked,
     this.onPopInvokedWithResult,
     required this.child,
   });
 
   @override
   Widget build(BuildContext context) {
-    if (canPop == null &&
-        onPopInvoked == null &&
-        onPopInvokedWithResult == null) {
+    if (canPop == null && onPopInvokedWithResult == null) {
       return child;
     }
 
@@ -39,26 +33,14 @@ class ReactiveFormPopScope<T> extends StatelessWidget {
       builder: (context, formGroup, _) {
         return PopScope<T>(
           canPop: canPop?.call(formGroup) ?? true,
-          onPopInvokedWithResult: _buildOnPopInvokedCallback(formGroup),
+          onPopInvokedWithResult:
+              onPopInvokedWithResult != null
+                  ? (didPop, result) =>
+                      onPopInvokedWithResult!(formGroup, didPop, result)
+                  : null,
           child: child,
         );
       },
     );
-  }
-
-  /// Builds the onPopInvoked callback based on the available callbacks.
-  PopInvokedWithResultCallback<T>? _buildOnPopInvokedCallback(
-    FormGroup formGroup,
-  ) {
-    if (onPopInvokedWithResult != null) {
-      return (didPop, result) =>
-          onPopInvokedWithResult!(formGroup, didPop, result);
-    }
-
-    if (onPopInvoked != null) {
-      return (didPop, _) => onPopInvoked!(formGroup, didPop);
-    }
-
-    return null;
   }
 }